Cleaned, participant-level data

Participant-level data for the Christmas statistical cognition experiment

Details

duration: Qualtrics-reported time taken in the experiment (seconds)

id: participant id

consent: Response to informed consent

times: "Have you participated previously"?

shuffle: "Do you understand use of shuffle reports?"

shuffle_other" (text) response for "Other" responses to shuffle

response: primary response: which team is faster?

confidence: participant's confidence in their response

salient_factors: (text) "What facts or observations were most salient in coming to the conclusion that you did?"

expt_strategy: (text) "Please describe your experimental strategy, if any, in your own words."

shuffle_desc: (text) "Did you make use of the "random shuffle reports"? If so, how?"

is_science: "Is your work in a field that would typically be considered scientific?"

is_science_other: (text) response for "Other" responses to is_science

education: highest level of formal education in scientific field

formal_training: years of formal statistical training

how_use: "How do statistics play a role in your work?" Concatenated string of responses

field: "In what applied field(s) do you use statistics?" Concatenated string of responses

preferred: "What sort of inferential procedures would you typically prefer?"

preferred_other: (text) response for "Other" responses to preferred

sig_testing: "What is your opinion about statistical significance testing?" Concatenated string of responses

mobile: Qualtrics flag for mobile browsers. Should be blank for all participants

evidence_power: Randomly-assigned transformation power. 3 or 7

effect_size: "true" effect size in standard deviation units, hidden from the participant

finished_practice_timestamp: javascript timestamp reported by browser when participant ended the instructions. Indexed to initial_timestamp

downloaded_expt: Did the participant download their experimental samples?

downloaded_null: Did the participant download their null samples?

initial_timestamp: initial timestamp (should be 0, start of experiment)

true_winner: Which "team" was the true winner? Corresponds to sign of effect_size

true_null: Was the effect size 0?

response_null: Was the participant's response "null"? (same or no detection)

response_alt: Negation of response_null

n_expt: Number of experimental samples requested

n_null: Number of random shuffle reports requested

how_use.analysis: (logical) Did the participant select the corresponding response to how_use?

how_use.develop: (logical) Did the participant select the corresponding response to how_use?

how_use.philosophy: (logical) Did the participant select the corresponding response to how_use?

how_use.method_comment: (logical) Did the participant select the corresponding response to how_use?

how_use.none: (logical) Did the participant select the corresponding response to how_use?

how_use.other: (logical) Did the participant select the corresponding response to how_use?

field.bio: (logical) Did the participant select the corresponding response to field?

field.phys: (logical) Did the participant select the corresponding response to field?

field.soc_beh: (logical) Did the participant select the corresponding response to field?

field.comp_tech: (logical) Did the participant select the corresponding response to field?

field.med: (logical) Did the participant select the corresponding response to field?

field.other: (text) If applicable, text response to "Other" responses for field.

sig_testing.necessary: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.prefer_other: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.misunderstood: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.no_opinion: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.discontinued: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.fatally_flawed: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.do_not_understand: (logical) Did the participant select the corresponding response to sig_testing?

sig_testing.other: (text) If applicable, text response to "Other" responses to sig_testing

text_comparison: (Strategy questions) Response mentioned comparison to the random shuffle reports

text_asymmetry: (Strategy questions) Response mentioned symmetry/asymmetry in the experimental samples

text_sampling_var: (Strategy questions) Response mentioned random shuffle reports as a means to assess sampling variability, distribution under the null, chance distribution, etc

text_inc_asymmetry: (Strategy questions) Response mentioned increasing asymmetry (or lack thereof) as sample size increased

text_no_shuffle: (Strategy questions) Explicitly said they did not use the shuffle reports

text_irrelevant: (Strategy questions) Text was irrelevant

text_missing: (Strategy questions) Text was missing