Generate image to show the interface with p values overlaid

Generate image to show the interface with p values overlaid

fig_evidence_p_vals(
  q,
  ps = c(0.005, 0.025, 0.05, 0.125, 0.25, 0.5, 0.75, 0.875, 0.95, 0.975, 0.995),
  letter = ""
)

Arguments

q
Evidence scale/power

ps
vector of p values to plot

letter
label to add to plot

Value

Returns a file path to the generated PNG file

Examples

fig_evidence_p_vals(q = 7)

#> [1] "/var/folders/yn/wzkl_t8567l9pvwhp0wsm40w0000gn/T//Rtmp939wSC/filedbf862016c82.png"