Coding of the text responses

Categorization of each participants' strategy-related free-text responses

Details

id: participant id

text_comparison: Response mentioned comparison to the random shuffle reports

text_asymmetry: Response mentioned symmetry/asymmetry in the experimental samples

text_sampling_var: Response mentioned random shuffle reports as a means to assess sampling variability, distribution under the null, chance distribution, etc

text_inc_asymmetry: Response mentioned increasing asymmetry (or lack thereof) as sample size increased

text_no_shuffle: Explicitly said they did not use the shuffle reports

text_irrelevant: Text was irrelevant

text_missing: Text was missing